LibRuntime: Move arc4random_buf, ferror, and fwrite from LibC
