+++ Userland/Libraries/LibC/stdio.cpp
@@ -242,14 +242,6 @@ void clearerr(FILE* stream)
     stream->clear_err();
 }
 
-// https://pubs.opengroup.org/onlinepubs/9699919799/functions/ferror.html
-int ferror(FILE* stream)
-{
-    VERIFY(stream);
-    ScopedFileLock lock(stream);
-    return stream->error();
-}
-
 size_t fread_unlocked(void* ptr, size_t size, size_t nmemb, FILE* stream)
 {
     VERIFY(stream);
@@ -267,19 +259,6 @@ size_t fread(void* ptr, size_t size, size_t nmemb, FILE* stream)
     VERIFY(stream);
     ScopedFileLock lock(stream);
     return fread_unlocked(ptr, size, nmemb, stream);
-}
-
-// https://pubs.opengroup.org/onlinepubs/9699919799/functions/fwrite.html
-size_t fwrite(void const* ptr, size_t size, size_t nmemb, FILE* stream)
-{
-    VERIFY(stream);
-    VERIFY(!Checked<size_t>::multiplication_would_overflow(size, nmemb));
-
-    ScopedFileLock lock(stream);
-    size_t nwritten = stream->write(reinterpret_cast<u8 const*>(ptr), size * nmemb);
-    if (!nwritten)
-        return 0;
-    return nwritten / size;
 }
 
 // https://pubs.opengroup.org/onlinepubs/9699919799/functions/fseek.html
